Прибрался в пректе после бездумного копирования файлов из одного проекта в другой.
