+++ packages/frontend/package.json
@@ -1,25 +1,31 @@
 {
-    "name": "@kio/frontend",
-    "private": true,
-    "version": "0.0.0",
-    "type": "module",
-    "scripts": {
-        "start": "vite serve --force",
-        "build": "tsc && vite build",
-        "preview": "vite preview"
-    },
-    "dependencies": {
-        "react": "^18.2.0",
-        "react-dom": "^18.2.0",
-        "typescript-plugin-css-modules": "^4.2.3",
-        "axios": "~1.3.6"
-    },
-    "devDependencies": {
-        "@types/react": "^18.0.27",
-        "@types/react-dom": "^18.0.10",
-        "@vitejs/plugin-react-swc": "^3.0.0",
-        "typescript": "^4.9.3",
-        "vite": "^4.1.0",
-        "prettier": "2.8.4"
-    }
+  "name": "@kio/frontend",
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
+  "scripts": {
+    "start": "vite serve --force",
+    "build": "tsc && vite build",
+    "preview": "vite preview"
+  },
+  "dependencies": {
+    "axios": "~1.3.6",
+    "marked": "5.0.1",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-query": "3.39.3",
+    "react-router-dom": "6.11.1",
+    "react-simplemde-editor": "5.2.0",
+    "react-transition-group": "4.4.5"
+  },
+  "devDependencies": {
+    "@types/react": "18.0.27",
+    "@types/react-dom": "18.0.10",
+    "@types/marked": "4.3.0",
+    "@vitejs/plugin-react-swc": "3.0.0",
+    "prettier": "2.8.4",
+    "typescript": "4.9.3",
+    "typescript-plugin-css-modules": "4.2.3",
+    "vite": "4.1.0"
+  }
 }
