--- packages/frontend/src/app.component.tsx
@@ -1,14 +1,0 @@
-import React, { useEffect, useState } from 'react';
-import axios from 'axios';
-
-export const App: React.FC = () => {
-    const [cats, setCats] = useState(0);
-
-    useEffect(() => {
-        axios.get('http://localhost:3001/get-cats-amount').then(({ data }) => {
-            setCats(Number(data));
-        });
-    }, []);
-
-    return <div className="App">cats amount: {cats}</div>;
-};
